+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* docs: Make /docs deep-redirect to readthedocs, add link in help
 * install: Remove generated deploy.sh scripts, just use "make install" instead
 
 ## 2.0.0-beta3 (2018-11-27)